source code cleanup
